Add Category and Brand Controllers and GetAllQuery

+++ Persistance/Repository/Base/Repository.cs
@@ -21,7 +21,7 @@ namespace Persistance.Repository.Base
             _set = _context.Set<T>();
         }
 
-        public async Task<T> GetById(int id)
+        public async Task<T> GetById(long id)
         {
             return await _set.FindAsync(id);
         }
